Testing Ephemerides.jl

+++ src/spice/SpiceFunctions.jl
@@ -46,12 +46,12 @@ Return ephemerides
 - `eph::EphemerisProvider`: Ephemeris provider object
 - `initialEpochTime::Float64`: Initial epoch time [s]
 - `times::Vector{Float64}`: Times since initial epoch
-- `targetBodyID::Int64`: Target body SPICE ID for ephemerides
-- `observerBodyID::Int64`: Observer body SPICE ID for ephemerides
-- `referenceBodyID::Int64`: Body SPICE ID for ephemerides reference (common to both)
+- `targetBodyID::Int16`: Target body SPICE ID for ephemerides
+- `observerBodyID::Int16`: Observer body SPICE ID for ephemerides
+- `referenceBodyID::Int16`: Body SPICE ID for ephemerides reference (common to both)
 - `frame::String`: Reference frame (default = "ECLIPJ2000")
 """
-function getEphemerides(eph::Ephemerides.EphemerisProvider, initialEpochTime::Float64, times::Vector{Float64}, targetBodyID::Int64, observerBodyID::Int64, referenceBodyID::Int64; frame::String = "ECLIPJ2000")
+function getEphemerides(eph::Ephemerides.EphemerisProvider, initialEpochTime::Float64, times::Vector{Float64}, targetBodyID::Int16, observerBodyID::Int16, referenceBodyID::Int16; frame::String = "ECLIPJ2000")
     ephemerisTimes::Vector{Float64} = initialEpochTime .+ times
     if frame == "ECLIPJ2000"
         i::Float64 = 23.43929111*pi/180
